Fixed confirmExit for Electron apps.
The previous implementation did not work because the window was unloaded
before the dialog could be opened. Also, opening a new workspace in the
existing window would not have been blocked.

This implementation uses a beforeunload handler in the frontend, and
uses the returnValue property of the event to disable unload if
required.